+++ neuron/__main__.py
@@ -1,5 +1,5 @@
 import logging
-from neuron import server, settings
+from neuron import server, settings, factory, protocol
 
 
 def main():
@@ -11,9 +11,16 @@ def main():
         logging.basicConfig(level=logging.DEBUG)
     else:
         logging.basicConfig(level=logging.INFO)
+    
+    # Set up the websocket factory. This is the central router
+    # of the application, connecting the backend to the frontend clients
+    router = factory.WSManagerFactory()
+    router.protocol = protocol.WSProtocol
+    router.setProtocolOptions(**settings.AUTOBAHN_PROTOCOL_OPTIONS)
+
     # TODO: Set up installed plugins
 
-    server.run_server()
+    server.run_server(router)
 
 if __name__ == '__main__':
     main();
